requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.2
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
binance==0.3
blinker==1.8.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.0
dateparser==1.2.0
Flask==3.0.3
Flask-JWT-Extended==4.6.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
greenlet==3.0.3
hyperlink==21.0.0
idna==3.7
incremental==24.7.1
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pyOpenSSL==24.2.1
python-binance==0.7.9
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
regex==2024.7.24
requests==2.32.3
service-identity==24.1.0
setuptools==71.1.0
six==1.16.0
SQLAlchemy==2.0.31
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
yarl==1.9.4
zope.interface==6.4.post2
